Moved Interpolate and GPs to new sub-directory
