+++ src/SolveROMs.py
@@ -7,15 +7,17 @@ import  os;
 Physics_Path        : str   = os.path.abspath(os.path.join(os.path.dirname(__file__), "Physics"));
 LD_Path             : str   = os.path.abspath(os.path.join(os.path.dirname(__file__), "LatentDynamics"));
 EncoderDecoder_Path : str   = os.path.abspath(os.path.join(os.path.dirname(__file__), "EncoderDecoder"));
+Interpolate_Path    : str   = os.path.abspath(os.path.join(os.path.dirname(__file__), "Interpolate"));
 Utilities_Path      : str   = os.path.abspath(os.path.join(os.path.dirname(__file__), "Utilities"));
 sys.path.append(Physics_Path);
 sys.path.append(LD_Path);
 sys.path.append(EncoderDecoder_Path);
+sys.path.append(Interpolate_Path);
 sys.path.append(Utilities_Path);
 
 import  torch;
 import  numpy;
-from    Interpolate                 import  Interpolate;
+from   .Interpolate                 import  Interpolate;
 from    Physics                     import  Physics;
 from    LatentDynamics              import  LatentDynamics;
 from    ParameterSpace              import  ParameterSpace;
